refactor(cmd): move obtain and renew functions
